refactor: Add exception handling in check_code_view This commit refactors the check_code_view function. It adds a try-except block to handle the case when the CustomUser.DoesNotExist exception is raised. If the exception is caught, an error message is displayed and the user is redirected to the login page.
feat: Add LoginAPI for user login API.
This commit introduces the LoginAPI class-based view for handling user login through an API endpoint.
The post method validates the provided phone number using the LoginSerializer.
It attempts to authenticate the user and create a verification code if successful.
The view returns a response with the user_id if authentication is successful to use user_id in CheckCodeAPI

feat: Add CheckCodeAPI for code verification API.
This commit adds the CheckCodeAPI class-based view for handling code verification through an API endpoint.
The post method validates the user_id and code using the CodeVarifySerializer.
It retrieves the user and checks if the code matches and is not expired.
If successful, it generates an access token and refresh token using RefreshToken from djangorestframework-simplejwt library.