feat: Add custom permission IsOwner. To handle object-level permissions in Django REST Framework. The IsOwner permission checks if the authenticated user is the owner of the object being accessed. It overrides the has_object_permission method and returns True only if the object matches the request user. This permission can be used to restrict access to objects based on ownership and provides an additional layer of security and data protection.
feat: Add UserInfoAPI with custom permissions.
To handle retrieving user information in Django REST Framework. The view requires both authentication and ownership permissions, as specified by the IsAuthenticated and IsOwner permission classes.
The view implements a GET method, which retrieves user information based on the provided 'pk' query parameter. If the 'pk' is provided, the view attempts to fetch the corresponding CustomUser object. If the user does not exist, a response with an appropriate error message is returned. The view then checks the object permissions to ensure the requesting user is the owner of the user object.
If the permissions are successfully validated, the view serializes the user object using the UserSerializer and returns the serialized data in the response. If the 'pk' is not provided, a response with a message to send the 'pk' parameter is returned.

feat: Add EditUserInfoAPI with IsAuthenticated and IsOwner permissions.
The view handles POST requests to update user information based on the provided 'pk' query parameter. If a valid 'pk' is provided, the view attempts to fetch the corresponding CustomUser object. If the user does not exist, a response with a descriptive error message is returned. The view then checks the object permissions to ensure that the requesting user is the owner of the user object.
If the permissions are successfully validated, the view initializes the UpdateUserSerializer with the user object and the request data (in partial mode to allow partial updates). The serializer is then validated, and if valid, the view checks if at least one field has been provided for update. If no fields are provided, a response indicating that at least one field must be entered is returned.
If the serializer is valid and at least one field is provided, the view saves the serializer and returns a response indicating successful update.
If the serializer is invalid, the view returns a response with the serializer errors.