refactor: Add 'id' and 'slug' to fields in class Meta of CategorySerializer
feat: Improve category handling in create method of AdCreateOrUpdateSerializer

In this commit, the create method of AdCreateOrUpdateSerializer is enhanced to improve category handling. Instead of directly querying and validating categories in the create method, a new function named validate_categorise is introduced. This function is defined outside the serializer and is responsible for validating and retrieving Category objects based on provided identifiers (category names or primary keys).

The validate_categorise function iterates through the list of category identifiers (categories_inputs) and attempts to find the corresponding Category objects by name and then by primary key (pk). If neither name nor pk match any existing Category, it raises a serializers.ValidationError to indicate that the provided identifier is not valid or does not correspond to any existing Category.

Back in the create method, the validate_categorise function is now called to validate the provided category values and obtain a list of valid Category objects (categories_list). If any invalid or non-existent category identifiers are found, the create method will raise a serializers.ValidationError, preventing the creation of the Ad instance.

With the introduction of the validate_categorise function and its integration into the serializer's create method, the category handling has been separated and centralized, resulting in a more organized and reliable approach. The create method now focuses on creating the Ad instance with the validated data, and the validate_categorise function takes care of the category validation, promoting code reusability and maintainability. And still category can be emtpy in method create.