feat: Add IsAdOwner custom permission
The has_object_permission compares the request.user with the obj.author, which represents the ad's author. If they match, permission is granted, allowing the user to perform actions restricted to ad owners. Otherwise, access is denied.

This custom permission enhances the access control in your views or viewsets, ensuring that only the ad owners have permission to perform certain actions on their ads.

refactor: Update AdCreateSerializer to AdCreateOrUpdateSerializer

This commit refactors the AdCreateSerializer to AdCreateOrUpdateSerializer, which now override method update for handling updates to existing Ad instances.

In the updated serializer, the fields and structure remain mostly unchanged, still handling fields such as 'title', 'text', 'image', 'status_product', 'price', 'phone', 'location', 'category', and 'active'. The category field is optional, allowing it to be empty or omitted during both create and update actions.

The crucial addition to the AdCreateOrUpdateSerializer is the update method, which implements the logic required to update existing Ad instances. When the serializer is used to update an Ad instance, the validate_categorise function is called to validate and retrieve the relevant Category objects based on provided identifiers (name or pk), similar to the create operation.

During the update process, the existing instance.category is cleared, and the updated categories_list (if any) is associated with the Ad instance. Additionally, specific fields are updated using the validated data, and the instance is saved with the specified update_fields. Furthermore, the confirmation field is set to False to facilitate admin re-check, and the instance is saved with this change.

This refactoring optimizes the serializer, allowing it to handle both create and update operations efficiently. The AdCreateOrUpdateSerializer is now a comprehensive solution, streamlining the API's interactions with Ad objects and providing consistent and reliable data handling for both creation and modification scenarios.