feat: Add SearchSerializer
In this commit, the 'SearchSerializer' has been added. This serializer is responsible for validating search query parameters in the API view. It includes a single field 'q', which represents the search query string and is required for conducting search operations.

The 'SearchSerializer' ensures that the 'q' parameter is provided in the incoming request data and that it is a valid string. If the 'q' parameter is missing or invalid, the serializer will raise a validation error, which will be handled appropriately in the API view.

The 'SearchSerializer' is now available for use in the API view, enabling the application to perform searches based on user-provided query strings.

feat: Add SearchAdAPI for Ad Searching

This commit introduces the 'SearchAdAPI', which allows users to perform searches for ads based on their query input. The 'SearchAdAPI' is a custom API view that handles search requests and returns a list of relevant ads based on the search query.

The API view uses the 'SearchSerializer' to validate the search query passed in the request data. If the provided query is valid, the API view performs a search on the active ads in the database, looking for matches in the ad titles, texts, and category names (case-insensitive).

The search results are then serialized using the 'AdListSerializer', which presents the relevant ad information in a structured format. The API view returns the serialized data as a response with a status code of 200 (OK) when successful.

If the search query is missing or invalid, the 'SearchAdAPI' will return the validation errors in the response with a status code of 400 (Bad Request).

refactor: rename serializer_classes to serializer_class in UpdateAdAPI

feat: Add search_ads in urls.py

'/search/': Linked to the 'SearchAdAPI' view