feat: Introduce CategoryListAPI
This API view allows users to retrieve a list of all available categories. The view fetches the category data from the database and uses the 'CategorySerializer' to serialize it. The serialized data is then returned as a response with a status code of 200 (OK) when the request is successful.

feat: Introduce AdsListWithCategoryAPI

This API view enables users to retrieve a list of ads belonging to a specific category. The view takes a category primary key (pk) as input and fetches the corresponding category from the database. If the category with the provided pk does not exist, the view returns an error response with a status code of 400 (Bad Request). If the category exists, the view queries the active and confirmed ads associated with that category and uses the 'AdListSerializer' to serialize the ad data. The serialized data is then returned as a response with a status code of 200 (OK) when the request is successful.

feat: Add categories_list in urls.py

'/list/category/': Linked to the 'CategoryListAPI' view

feat: Add ads_list_with_category in urls.py

'/category/<int:pk>/': Linked to the 'AdsListWithCategoryAPI' view and takes an integer parameter (pk) to identify the specific Category instance for filtering ads