refactor: Update 'status_product' Field Choices
In this commit, the choices for the 'status_product' field have been refactored to use more descriptive and user-friendly values. The previous choices used numeric values (0, 1, 2, 3), which were less intuitive for users to understand. The new choices use descriptive text values, making it easier for users to select the appropriate status for their ads.

The new 'STATUS_CHOICES' for the 'status_product' field are as follows:

'need repair': This option is used when the product needs repair or maintenance.

'worked': This option indicates that the product is in working condition.

'like new': This option is selected for products that are in excellent condition, almost like new.

'new': This option is used for brand new products that have not been used before.

'max_length': from 10 change to 30

By updating the 'status_product' field choices, the application provides a more user-friendly interface for users to specify the status of their ads accurately. The previous numeric values have been replaced with descriptive text values, improving the overall user experience and clarity of the application's features.