refactor: Update 'price' Field in Ad Model
This commit refactors the 'price' field in the 'Ad' model to use a 'PositiveBigIntegerField' with appropriate validators. The 'price' field previously used a 'CharField' with a maximum length of 100, but it has been improved to enforce positive integer values within a specific range.

The new 'price' field now uses the 'PositiveBigIntegerField', which ensures that the price is represented as a positive integer greater than or equal to 10,000 and less than or equal to 99,999,999,999. This validation is achieved through the use of 'MinValueValidator' and 'MaxValueValidator' with appropriate minimum and maximum values.

By making this change, the 'price' field is more robust and provides better data integrity, ensuring that only valid and meaningful price values can be stored in the database.

refactor: Improve Validation Error Message in 'validate_categorise' Function

Now, the 'validate_categorise' function raises a 'serializers.ValidationError' with a more descriptive error message. It includes the field name 'category' as part of the error message, making it easier to pinpoint the exact field causing the validation error.